pyproject.toml
[project]
name = "angel-one-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "logzero>=1.7.0",
    "mcp[cli]>=1.9.2",
    "pyotp>=2.9.0",
    "requests>=2.32.3",
    "smartapi-python>=1.5.5",
    "websocket>=0.2.1",
    "websockets>=15.0.1",
]
